refactor(ast): do not generate node type aliases via a derive macro
This removes the NodeContent macro and replaces it with a macro_rules-based implementation.
Display-related code has been moved to the `NodeContentDisplay` derive macro.
